# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inferno SMP Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
}

export interface RegisterBody {
  username: string;
  password: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Rank {
  id: string;
  name: string;
  price: number;
  homes: number;
  rtpCooldown?: number | null;
  tag?: string | null;
  color: string;
}

export interface BuyBody {
  rank: string;
  transactionId: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Order {
  id: number;
  username: string;
  rank: string;
  transactionId: string;
  status: OrderStatus;
  createdAt: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateOrderStatusBody {
  orderId: number;
  status: UpdateOrderStatusBodyStatus;
}