# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inferno SMP Store API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.enum(["admin", "user"]),
  }),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
});

/**
 * @summary Get all store ranks
 */
export const GetStoreResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  price: zod.number(),
  homes: zod.number(),
  rtpCooldown: zod.number().nullish(),
  tag: zod.string().nullish(),
  color: zod.string(),
});
export const GetStoreResponse = zod.array(GetStoreResponseItem);

/**
 * @summary Purchase a rank
 */
export const BuyRankBody = zod.object({
  rank: zod.string(),
  transactionId: zod.string(),
});

/**
 * @summary Get current user's orders
 */
export const GetOrdersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  transactionId: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  createdAt: zod.string(),
});
export const GetOrdersResponse = zod.array(GetOrdersResponseItem);

/**
 * @summary Get all orders (admin only)
 */
export const GetAdminOrdersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  transactionId: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  createdAt: zod.string(),
});
export const GetAdminOrdersResponse = zod.array(GetAdminOrdersResponseItem);

/**
 * @summary Update order status (admin only)
 */
export const UpdateOrderStatusBody = zod.object({
  orderId: zod.number(),
  status: zod.enum(["pending", "approved", "rejected"]),
});

export const UpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  transactionId: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  createdAt: zod.string(),
});