# lib/api-zod/src/generated/types/buyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inferno SMP Store API
 * OpenAPI spec version: 0.1.0
 */

export interface BuyBody {
  rank: string;
  transactionId: string;
}