# lib/api-zod/src/generated/types/rank.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inferno SMP Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Rank {
  id: string;
  name: string;
  price: number;
  homes: number;
  rtpCooldown?: number | null;
  tag?: string | null;
  color: string;
}